api/requirements.txt
azure-core==1.35.0
azure-storage-blob==12.26.0
blinker==1.9.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
cryptography==45.0.6
Flask==3.1.1
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.3.2
pandas==2.3.1
pycparser==2.22
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
six==1.17.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
python-dotenv==1.1.0
gunicorn